# ai_test/backend/tests/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Web测试平台接口文档",
    "description": "FastAPI",
    "version": "0.0.1"
  },
  "paths": {
    "/api/users/register": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "用户注册",
        "operationId": "register_api_users_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "用户登录",
        "description": "登录的逻辑,实现用户名、手机号、邮箱均可以作为账号登录",
        "operationId": "login_api_users_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginForm": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 20,
            "minLength": 6,
            "title": "Username",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "maxLength": 18,
            "minLength": 6,
            "title": "Password",
            "description": "密码"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginForm"
      },
      "LoginSchema": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token",
            "description": "访问令牌"
          },
          "user": {
            "$ref": "#/components/schemas/UserInfoSchema"
          }
        },
        "type": "object",
        "required": [
          "token",
          "user"
        ],
        "title": "LoginSchema"
      },
...
      "RegisterForm": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 20,
            "minLength": 6,
            "title": "Username",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "maxLength": 18,
            "minLength": 6,
            "title": "Password",
            "description": "密码"
          },
          "password_confirm": {
            "type": "string",
            "maxLength": 18,
            "minLength": 6,
            "title": "Password Confirm",
            "description": "确认密码"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "邮箱"
          },
          "mobile": {
            "type": "string",
            "title": "Mobile",
            "description": "手机号"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname",
            "description": "昵称",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "username",
          "password",
          "password_confirm",
          "email",
          "mobile"
        ],
        "title": "RegisterForm"
      },
...
      "UserInfoSchema": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "用户id"
          },
          "username": {
            "type": "string",
            "title": "Username",
            "description": "用户名"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname",
            "description": "昵称"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "邮箱"
          },
          "mobile": {
            "type": "string",
            "title": "Mobile",
            "description": "手机号"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "是否激活"
          },
          "is_superuser": {
            "type": "boolean",
            "title": "Is Superuser",
            "description": "是否超级管理员"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "nickname",
          "email",
          "mobile",
          "is_active",
          "is_superuser"
        ],
        "title": "UserInfoSchema"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/users/token"
          }
        }
      }
    }
  }
}